+++ src/main/java/net/minecraft/src/GuiSlotServer.java
@@ -220,7 +220,7 @@ class GuiSlotServer extends GuiSlot {
             }
         } else {
             var15 = 1;
-            var16 = (int) (Minecraft.getSystemTime() / 100L + (long) (par1 * 2L) & 7L);
+            var16 = (int) (Minecraft.getSystemTime() / 100L + (par1 * 2L) & 7L);
 
             if (var16 > 4) {
                 var16 = 8 - var16;
